Move api/server/router/network to daemon/server/router/network
